+++ lib/Payever/ExternalIntegration/Core/Engine.php
@@ -15,9 +15,9 @@
 namespace Payever\ExternalIntegration\Core;
 
 // @codeCoverageIgnoreStart
-define('PEI_CORE_VERSION', '2.4.0');
+define('PEI_CORE_VERSION', '2.5.0');
 define('PEI_CORE_MAJOR_VERSION', 2);
-define('PEI_CORE_MINOR_VERSION', 4);
+define('PEI_CORE_MINOR_VERSION', 5);
 define('PEI_CORE_RELEASE_VERSION', 0);
 
 define('PEI_NAMESPACE', 'Payever\ExternalIntegration');