+++ lib/Payever/ExternalIntegration/Products/Http/MessageEntity/ProductVariantOptionEntity.php
@@ -1,0 +1,36 @@
+<?php
+/**
+ * PHP version 5.4 and 7
+ *
+ * @package   Payever\Products
+ * @author    Hennadii.Shymanskyi <gendosua@gmail.com>
+ * @copyright 2017-2019 payever GmbH
+ * @license   MIT <https://opensource.org/licenses/MIT>
+ */
+
+namespace Payever\ExternalIntegration\Products\Http\MessageEntity;
+
+use Payever\ExternalIntegration\Core\Base\MessageEntity;
+
+/**
+ * PHP version 5.4 and 7
+ *
+ * @package   Payever\Products
+ * @author    payever GmbH <service@payever.de>
+ * @author    Hennadii.Shymanskyi <gendosua@gmail.com>
+ * @copyright 2017-2019 payever GmbH
+ * @license   MIT <https://opensource.org/licenses/MIT>
+ *
+ * @method string getName()
+ * @method string getValue()
+ * @method self setName(string $name)
+ * @method self setValue(string $value)
+ */
+class ProductVariantOptionEntity extends MessageEntity
+{
+    /** @var string */
+    protected $name;
+
+    /** @var string */
+    protected $value;
+}