+++ lib/Payever/ExternalIntegration/ThirdParty/Enum/DirectionEnum.php
@@ -20,13 +20,8 @@ use Payever\ExternalIntegration\Core\Base\EnumerableConstants;
  * @copyright 2017-2019 payever GmbH
  * @license   MIT <https://opensource.org/licenses/MIT>
  */
-class Action extends EnumerableConstants
+class DirectionEnum extends EnumerableConstants
 {
-    const ACTION_CREATE_PRODUCT      = 'create-product';
-    const ACTION_UPDATE_PRODUCT      = 'update-product';
-    const ACTION_REMOVE_PRODUCT      = 'remove-product';
-    const ACTION_ADD_INVENTORY       = 'add-inventory';
-    const ACTION_SET_INVENTORY       = 'set-inventory';
-    const ACTION_SUBTRACT_INVENTORY  = 'subtract-inventory';
-    const ACTION_PRODUCTS_SYNC       = 'products-sync';
+    const INWARD = 'inward';
+    const OUTWARD = 'outward';
 }