+++ tests/Unit/ExternalIntegration/Products/Http/RequestEntity/ProductRequestEntityTest.php
@@ -5,13 +5,14 @@ namespace Payever\Tests\Unit\ExternalIntegration\Products\Http\RequestEntity;
 use Payever\ExternalIntegration\Products\Http\RequestEntity\ProductRequestEntity;
 use Payever\Tests\Unit\ExternalIntegration\Core\Http\AbstractMessageEntityTest;
 use Payever\Tests\Unit\ExternalIntegration\Products\Http\MessageEntity\ProductCategoryEntityTest;
+use Payever\Tests\Unit\ExternalIntegration\Products\Http\MessageEntity\ProductVariantOptionEntityTest;
 
 class ProductRequestEntityTest extends AbstractMessageEntityTest
 {
     protected static $scheme = array(
         'externalId' => 'stub_id',
         'imagesUrl' => array('https://some.com/img.png'),
-        'enabled' => false,
+        'active' => true,
         'uuid' => 'stub_uuid',
         'businessUuid' => 'stub_business_uuid',
         'categories' => array(),
@@ -20,14 +21,20 @@ class ProductRequestEntityTest extends AbstractMessageEntityTest
         'description' => 'product description',
         'price' => 111.1,
         'salePrice' => 101.1,
-        'hidden' => false,
+        'onSales' => true,
         'sku' => 'stub_sku',
         'type' => 'physical',
+        'variants' => array(),
     );
 
     public static function getScheme()
     {
         $scheme = parent::getScheme();
+
+        $variant = $scheme;
+        $variant['options'] = array(ProductVariantOptionEntityTest::getScheme());
+
+        $scheme['variants'][] = $variant;
         $scheme['categories'][] = ProductCategoryEntityTest::getScheme();
 
         return $scheme;