+++ tests/Unit/ExternalIntegration/ThirdParty/Action/InwardActionProcessorTest.php
@@ -4,17 +4,17 @@ namespace Payever\Tests\Unit\ExternalIntegration\ThirdParty\Action;
 
 use Payever\ExternalIntegration\Core\Logger\NullLogger;
 use Payever\ExternalIntegration\ThirdParty\Action\ActionHandlerPool;
-use Payever\ExternalIntegration\ThirdParty\Action\ActionRequestProcessor;
 use Payever\ExternalIntegration\ThirdParty\Action\ActionResult;
-use Payever\ExternalIntegration\ThirdParty\Enum\Action;
+use Payever\ExternalIntegration\ThirdParty\Action\InwardActionProcessor;
+use Payever\ExternalIntegration\ThirdParty\Enum\ActionEnum;
 use Payever\Tests\Bootstrap\TestCase;
 
-class ActionRequestProcessorTest extends TestCase
+class InwardActionProcessorTest extends TestCase
 {
     /** @var ActionHandlerPool */
     private $handlerPool;
 
-    /** @var ActionRequestProcessor */
+    /** @var InwardActionProcessor */
     private $actionRequestProcessor;
 
     protected function setUp()
@@ -23,7 +23,7 @@ class ActionRequestProcessorTest extends TestCase
         $actionResult = new ActionResult();
         $logger = new NullLogger();
 
-        $this->actionRequestProcessor = new ActionRequestProcessor(
+        $this->actionRequestProcessor = new InwardActionProcessor(
             $this->handlerPool,
             $actionResult,
             $logger
@@ -35,19 +35,28 @@ class ActionRequestProcessorTest extends TestCase
      */
     public function testNoHandlers()
     {
-        $this->actionRequestProcessor->processActionRequest('stub');
+        $this->actionRequestProcessor->process('stub');
     }
 
-    public function testSuccess()
+    /**
+     * @dataProvider actionListDataProvider
+     *
+     * @param string $action
+     * @throws \Exception
+     */
+    public function testSuccess($action)
     {
-        $action = Action::ACTION_CREATE_PRODUCT;
-
         $handler = $this->getMockForAbstractClass('Payever\ExternalIntegration\ThirdParty\Action\ActionHandlerInterface');
         $handler->expects($this->once())->method('getSupportedAction')->willReturn($action);
         $handler->expects($this->once())->method('handle')->withAnyParameters();
 
         $this->handlerPool->registerActionHandler($handler);
 
-        $this->actionRequestProcessor->processActionRequest($action, '{}');
+        $this->actionRequestProcessor->process($action, '{}');
+    }
+
+    public function actionListDataProvider()
+    {
+        return array_map(function ($el) { return [$el]; }, ActionEnum::enum());
     }
 }